requirements.txt
numpy>=1.22.2
tqdm>=4.64.0
opencv-contrib-python>=4.5.5
matplotlib>=3.2.0
scikit-learn>=1.1.1
einops>=0.3.2
scipy>=1.7.3
seaborn>=0.11.0
SimpleITK
pyradiomics
tensorflow-cpu>=2.7.0
